Fix logic bugs.
scores changes
 * lines 57 - 122, work around because it appears Maud doesn't let you
   have mutable variables. Was particularly banging my head against the
desk for line 100... thanks chatgpt.
 * fn group_by_scores: using a HashMap was messing up data ordering, but
   it's also probably a fast way to do this fn, so I left it as is (not
that perf really matters here but still!), I just added a sort_scores fn
to resort once when done.
 * fn group_by_bettor_name_and_round was returning all kinds of
   mis-sorted data, for same reason (HashMap). Took a while to step
through that code and fix it up, but I learned about BTreeMap as a
silver lining.

+++ src/model.rs
@@ -74,8 +74,8 @@ pub struct ScoreData {
 
 pub struct SummaryScore {
 	pub bettor_name   :String,
-	pub computed_rounds :Vec<i64>,
-	pub new_scores      :Vec<i64>,
+	pub computed_rounds :Vec<isize>,
+	pub new_scores      :Vec<isize>,
 }
 
 pub struct  SummaryScores {
